rename : postSkeleton 위치 이동
